moving over components to a proper folder
