--- app/credentials/_components/index.ts
@@ -1,1 +1,0 @@
-export { default as CertContainer } from "./CertContainer";
