+++ components/mainPageComponents/ScreenLight.tsx
