+++ components/mainPageComponents/index.ts
@@ -1,6 +1,6 @@
 export { default as Experience } from "./Experience";
 export { default as Laptop } from "./Laptop";
 export { default as ScreenLight } from "./ScreenLight";
-export { LaptopScreen } from "./_screenComponents";
+export { LaptopScreen } from "./screen";
 export { default as ReactIcon } from "./ReactIcon";
 export { default as NodeIcon } from "./NodeIcon";
