+++ components/mainPageComponents/screen/Icon.tsx
