+++ components/mainPageComponents/screen/Sherlock.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { useLaptopInfo } from "../../_providers/LaptopInfoProvider";
+import { useLaptopInfo } from "@/components/providers/LaptopInfoProvider";
 import Draggable from "react-draggable";
 import { TechStack, WindowBar } from ".";
