+++ components/mainPageComponents/screen/TechIcon.tsx
