+++ components/mainPageComponents/screen/TechStack.tsx
@@ -1,4 +1,4 @@
-import { useLaptopInfo } from "../../_providers/LaptopInfoProvider";
+import { useLaptopInfo } from "@/components/providers/LaptopInfoProvider";
 import { TechIcon } from ".";
 
 const TechStack = () => {
