+++ components/mainPageComponents/screen/WindowBar.tsx
