+++ components/mainPageComponents/screen/index.ts
