+++ components/providers/LaptopInfoProvider.tsx
